Add enum Shape, encapsulating Sphere
Change other structs to use a Shape, instead of a Sphere.

Provide common functions in Shape, calling the same function
in the Sphere struct. There is perhaps a better way using
traits, but that route seems to bring other difficulties, as
a trait is not a concrete data type that can easily be passed
around and stored for later use.